requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pillow>=9.0.0
imagehash>=4.3.0
requests>=2.28.0
python-multipart>=0.0.5
firebase-admin>=6.0.0
transformers>=4.30.0
protobuf>=4.21.0,<4.22.0
torch>=2.0.0